issue-06: new IO vignette
This commit gives a new IO vignette, incorporating
the vignette from #10 as an exercise. The rmd from
#10 was moved to the hwsolutions folder as a
reference to instructors. It contains extensive use
of dplyr, so it may not be appropriate for students.

The crime.xls and crime.csv files were added to
have a static copy, shorten the urls, and provide
the students the "goal" for the exercise.

More could be done to flesh out saving R objects.

The section about connections could easily be
skipped.